+++ src/ai/index.js
@@ -1,6 +1,4 @@
 const distill = require('./distill');
 const interview = require('./interview');
-const feynman = require('./feynman');
-const testlab = require('./testlab');
 
-module.exports = { distill, interview, feynman, testlab };
+module.exports = { distill, interview };